Move LastLoginByUserIdQuery and handler interface
A quote from Eric Evans, teaching me why these classes belong in the
Application layer:

> Application Layer: Defines the jobs the software is supposed to do and
> directs the expressive domain objects to work out problems. The tasks
> this layer is responsible for are meaningful to the business or
> necessary for interaction with the application layers of other systems
